refactor: Further organize components
